refactor rollout weight update flow
