+++ xtuner/v1/rl/trainer/worker.py
@@ -48,6 +48,7 @@ from xtuner.v1.model.utils.misc import ModelForwardExtraLogInfo
 from xtuner.v1.profiler import profiling_memory, profiling_time
 from xtuner.v1.rl.loss import BaseRLLossConfig, BaseRLLossContext, finalize_train_policy_metrics, kl_penalty
 from xtuner.v1.rl.utils import SingleAcceleratorWorker
+from xtuner.v1.rl.weight_update import TrainRolloutMode, UpdateWeighter
 from xtuner.v1.train.trainer import LoadCheckpointConfig
 from xtuner.v1.utils import (
     XTUNER_DETERMINISTIC,
@@ -60,7 +61,6 @@ from xtuner.v1.utils import (
 )
 
 from ..rollout_is import merge_rollout_is_metrics
-from .update_weighter import UpdateWeighter
 
 
 DeviceMeshRaw: TypeAlias = List[List[int]]  # A list of lists representing device mesh indices
@@ -201,7 +201,7 @@ class WorkerLogItem(TypedDict):
     sft_train_metrics: NotRequired[dict[str, float]]
 
 
-class TrainingWorker(SingleAcceleratorWorker, UpdateWeighter):
+class TrainingWorker(SingleAcceleratorWorker):
     _SAVE_WEIGHTS_DIR = "weights"
     _SAVE_SFT_DATALOADER_DIR = "sft_dataloader"
     _SAVE_SFT_TRAIN_STATE_PATH = "sft_train_state.json"
@@ -269,7 +269,24 @@ class TrainingWorker(SingleAcceleratorWorker, UpdateWeighter):
             if hasattr(worker_cfg.model_cfg.text_config, "mtp_config"):
                 self.mtp_config = worker_cfg.model_cfg.text_config.mtp_config
 
-        self._init_update_weighter()
+        self.update_weighter = UpdateWeighter(
+            rank=self.rank,
+            logger=self.logger,
+            config=self.config,
+            engine=self._engine,
+        )
+
+    @ray_method
+    def update_rollout_info(self, *args, **kwargs):
+        return self.update_weighter.update_rollout_info(*args, **kwargs)
+
+    @ray_method
+    def set_train_rollout_mode(self, train_rollout_mode: TrainRolloutMode):
+        return self.update_weighter.set_train_rollout_mode(train_rollout_mode)
+
+    @ray_method
+    def update_weights(self):
+        return self.update_weighter.update_weights()
 
     def _init_sft(self, worker_cfg: WorkerConfig):
         self._sft_dataloader_config = worker_cfg.sft_dataloader_cfg
